+++ src/main/java/com/ajou/hertz/domain/instrument/effector/strategy/EffectorCreationStrategy.java
@@ -1,7 +1,8 @@
-package com.ajou.hertz.domain.instrument.strategy;
+package com.ajou.hertz.domain.instrument.effector.strategy;
 
-import com.ajou.hertz.domain.instrument.dto.request.CreateNewEffectorRequest;
-import com.ajou.hertz.domain.instrument.entity.Effector;
+import com.ajou.hertz.domain.instrument.effector.dto.request.CreateNewEffectorRequest;
+import com.ajou.hertz.domain.instrument.effector.entity.Effector;
+import com.ajou.hertz.domain.instrument.strategy.InstrumentCreationStrategy;
 import com.ajou.hertz.domain.user.entity.User;
 
 public class EffectorCreationStrategy implements InstrumentCreationStrategy<Effector, CreateNewEffectorRequest> {
